Identify the possible entity sets, their attributes, and relationships
From the following problem statement identify the possible entity sets, their attributes, and relationships.
SE VLabs Inc. is a young company with a few departments spread across the country. As of now, the company has a strength of 200+ employees.
Each employee works in a department. While joining, a person has to provide a lot of personal and professional details including name, address, phone #, mail address, date of birth, and so on. Once all these information are furnished, a unique ID is generated for each employee. He is then assigned a department in which he will work.
There are around ten departments in the company. Unfortunately, two departments were given same names. However, departments too have ID's, which are unique.
Note: Try to use the features of the interface provided to capture as much details as possible.